chore: Add YAML to list of supported languages
# Description

YAML is not on the list, however it is in fact supported. I have incorporates this library into my blog and setting `yaml` as the value for `language` in the `CopyBlock` works as desired.

## Demonstration 

You can find an example of YAML working happily on my blog [(with an example post here)](https://blog.dennisokeeffe.com/blog/2019-04-22-intro-to-helm) where I have gladly incorporated this package.